client/web: move session logic to auth.go
Updates tailscale/corp#14335
